[#659] Split NotesPage into smaller modules
- Extract NotebookFormDialog to separate file (~120 lines)
- Extract ShareDialogWrapper to separate file (~50 lines)
- Extract NoteHistoryPanel to separate file (~85 lines)
- Extract ViewState, DialogState types and transform functions
- Reduce NotesPage.tsx from 908 to 612 lines (~33% reduction)

Part of Epic #338
