+++ src/ui/pages/notes/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Notes page components and utilities.
+ * Part of Epic #338, Issue #659 (component splitting).
+ */
+export { NotebookFormDialog } from './NotebookFormDialog';
+export { ShareDialogWrapper } from './ShareDialogWrapper';
+export { NoteHistoryPanel } from './NoteHistoryPanel';
+export { toUINote, toUINotebook } from './types';
+export type { ViewState, DialogState } from './types';
